Refactor to add Live Edge detection use case added new activity for CameraX revised main activity in preperation for buttons to start diferent use cases for testing added support for https://github.com/hannesa2/LiveEdgeDetection
