+++ app/build.gradle
@@ -33,6 +33,7 @@ android {
     }
 }
 
+
 dependencies {
 
 //    Room DB Dependencies
@@ -57,6 +58,8 @@ dependencies {
     implementation "androidx.camera:camera-lifecycle:$camerax_version"
 // CameraX View class
     implementation "androidx.camera:camera-view:1.0.0-alpha14"
+
+    implementation "com.github.hannesa2:LiveEdgeDetection:1.8.5"
 
 //    default deps
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
