+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.CameraXPractice">
-        <activity android:name=".pictureViewEdit"></activity>
+        <activity android:name=".CameraX"></activity>
+        <activity android:name=".PictureViewEdit" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
