+++ app/src/main/java/com/example/cameraxpractice/MainActivity.kt
@@ -1,28 +1,21 @@
 package com.example.cameraxpractice
 
 import android.Manifest
-import android.annotation.SuppressLint
 import android.content.Intent
 import android.content.pm.PackageManager
-import android.graphics.Matrix
-import android.media.Image
 import android.net.Uri
 import android.os.Bundle
-import android.os.Environment
-import android.provider.MediaStore
 import android.util.Log
-import android.view.View
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.camera.core.*
 import androidx.camera.lifecycle.ProcessCameraProvider
-import androidx.camera.view.PreviewView
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
 import com.example.cameraxpractice.R.string.app_name
+import kotlinx.android.synthetic.main.activity_camera_x.*
 import kotlinx.android.synthetic.main.activity_main.*
 import java.io.File
-import java.nio.ByteBuffer
 import java.text.SimpleDateFormat
 import java.util.*
 import java.util.concurrent.ExecutorService
@@ -32,226 +25,12 @@ typealias ImageListener = (imageProxy: ImageProxy) -> Unit
 
 
 class MainActivity : AppCompatActivity() {
-    private var imageCapture: ImageCapture? = null
 
-    private lateinit var outputDirectory: File
-    private lateinit var cameraExecutor: ExecutorService
-
-    private class LuminosityAnalyzer(private val listener: ImageListener ) : ImageAnalysis.Analyzer {
-
-        private fun ByteBuffer.toByteArray(): ByteArray {
-            rewind()    // Rewind the buffer to zero
-            val data = ByteArray(remaining())
-            get(data)   // Copy the buffer into a byte array
-            return data // Return the byte array
-        }
-
-        override fun analyze(image: ImageProxy) {
-
-            val buffer = image.planes[0].buffer
-            val data = buffer.toByteArray()
-            val pixels = data.map { it.toInt() and 0xFF }
-//            val luma = pixels.average()
-
-            listener(image)
-
-            image.close()
-        }
-    }
-
-
-//    Activity actions
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        // Request camera permissions
-        if (allPermissionsGranted()) {
-            startCamera()
-        } else {
-            ActivityCompat.requestPermissions(
-                    this, REQUIRED_PERMISSIONS, REQUEST_CODE_PERMISSIONS)
-        }
-
-        // Set up the listener for take photo button
-        camera_capture_button.setOnClickListener { takePhoto() }
-
-        outputDirectory = getOutputDirectory()
-
-        cameraExecutor = Executors.newSingleThreadExecutor()
-    }
-
-    private fun takePhoto() {
-        // Get a stable reference of the modifiable image capture use case
-        val imageCapture = imageCapture ?: return
-
-        // Create time-stamped output file to hold the image
-        val photoFile = File(
-                outputDirectory,
-                SimpleDateFormat(FILENAME_FORMAT, Locale.US
-                ).format(System.currentTimeMillis()) + ".jpg")
-
-        // Create output options object which contains file + metadata
-        val outputOptions = ImageCapture.OutputFileOptions.Builder(photoFile).build()
-
-        // Set up image capture listener, which is triggered after photo has
-        // been taken
-        imageCapture.takePicture(
-                outputOptions, ContextCompat.getMainExecutor(this), object : ImageCapture.OnImageSavedCallback {
-            override fun onError(exc: ImageCaptureException) {
-                Log.e(TAG, "Photo capture failed: ${exc.message}", exc)
-            }
-
-            override fun onImageSaved(output: ImageCapture.OutputFileResults) {
-                val savedUri = Uri.fromFile(photoFile)
-                val msg = "Please Edit"
-                Toast.makeText(baseContext, msg, Toast.LENGTH_SHORT).show()
-                Log.d(TAG, msg)
-                goToViewEdit(savedUri)
-            }
-        })
     }
 
 
-    private fun startCamera() {
-        val cameraProviderFuture = ProcessCameraProvider.getInstance(this)
-        val imgProxy: ImageProxy
-
-        cameraProviderFuture.addListener(Runnable {
-            // Used to bind the lifecycle of cameras to the lifecycle owner
-            val cameraProvider: ProcessCameraProvider = cameraProviderFuture.get()
-
-            // Preview
-            val preview = Preview.Builder()
-                .build()
-                .also {
-                    it.setSurfaceProvider(viewFinder.createSurfaceProvider())
-                }
-            imageCapture = ImageCapture.Builder()
-                .build()
-
-//set to return matrix
-            val imageAnalyzer = ImageAnalysis.Builder()
-                .build()
-                .also {
-                    it.setAnalyzer(cameraExecutor, LuminosityAnalyzer { img ->
-//                            Log.d(TAG, "Average luminosity: $luma")
-                        getCorrectionMatrix(img , viewFinder)
-                    })
-                }
-
-            // Select back camera as a default
-            val cameraSelector = CameraSelector.DEFAULT_BACK_CAMERA
-
-            try {
-                // Unbind use cases before rebinding
-                cameraProvider.unbindAll()
-
-                // Bind use cases to camera
-                cameraProvider.bindToLifecycle(
-                        this, cameraSelector, preview, imageCapture, imageAnalyzer )
-
-
-            } catch (exc: Exception) {
-                Log.e(TAG, "Use case binding failed", exc)
-            }
-
-        }, ContextCompat.getMainExecutor(this))
-    }
-
-    private fun allPermissionsGranted() = REQUIRED_PERMISSIONS.all {
-        ContextCompat.checkSelfPermission(
-                baseContext, it) == PackageManager.PERMISSION_GRANTED
-    }
-// figure out how to store in DCIM
-    private fun getOutputDirectory(): File {
-        val mediaDir = externalMediaDirs.firstOrNull()?.let {
-            File(it, resources.getString(app_name)).apply { mkdirs() } }
-        return if (mediaDir != null && mediaDir.exists())
-            mediaDir else filesDir
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        cameraExecutor.shutdown()
-    }
-
-    companion object {
-        private const val TAG = "CameraXBasic"
-        private const val FILENAME_FORMAT = "yyyy-MM-dd-HH-mm-ss-SSS"
-        private const val REQUEST_CODE_PERMISSIONS = 10
-        private val REQUIRED_PERMISSIONS = arrayOf(Manifest.permission.WRITE_EXTERNAL_STORAGE, Manifest.permission.CAMERA)
-    }
-
-    override fun onRequestPermissionsResult(
-            requestCode: Int, permissions: Array<String>, grantResults:
-            IntArray) {
-        if (requestCode == REQUEST_CODE_PERMISSIONS) {
-            if (allPermissionsGranted()) {
-                startCamera()
-            } else {
-                Toast.makeText(this,
-                        "Permissions not granted by the user.",
-                        Toast.LENGTH_SHORT).show()
-                finish()
-            }
-        }
-    }
-
-
-
-
-//for learning about transforms still need to learn where to pass the matrix to in order to manipulate it
-    fun getCorrectionMatrix(imageProxy: ImageProxy, previewView: PreviewView) : Matrix {
-        val cropRect = imageProxy.cropRect
-        val rotationDegrees = imageProxy.imageInfo.rotationDegrees
-        val matrix = Matrix()
-
-        // A float array of the source vertices (crop rect) in clockwise order.
-        val source = floatArrayOf(
-            cropRect.left.toFloat(),
-            cropRect.top.toFloat(),
-            cropRect.right.toFloat(),
-            cropRect.top.toFloat(),
-            cropRect.right.toFloat(),
-            cropRect.bottom.toFloat(),
-            cropRect.left.toFloat(),
-            cropRect.bottom.toFloat()
-        )
-
-        // A float array of the destination vertices in clockwise order.
-        val destination = floatArrayOf(
-            0f,
-            0f,
-            previewView.width.toFloat(),
-            0f,
-            previewView.width.toFloat(),
-            previewView.height.toFloat(),
-            0f,
-            previewView.height.toFloat()
-        )
-
-        // The destination vertexes need to be shifted based on rotation degrees. The
-        // rotation degree represents the clockwise rotation needed to correct the image.
-
-        // Each vertex is represented by 2 float numbers in the vertices array.
-        val vertexSize = 2
-        // The destination needs to be shifted 1 vertex for every 90° rotation.
-        val shiftOffset = rotationDegrees / 90 * vertexSize;
-        val tempArray = destination.clone()
-        for (toIndex in source.indices) {
-            val fromIndex = (toIndex + shiftOffset) % source.size
-            destination[toIndex] = tempArray[fromIndex]
-        }
-        matrix.setPolyToPoly(source, 0, destination, 0, 4)
-        return matrix
-    }
-
-
-    fun goToViewEdit( uri: Uri ) {
-//        pass the URI for the picture to the ViewEdit
-        val intent = Intent(this, pictureViewEdit::class.java).apply {
-            putExtra("image url", uri.toString())
-        }
-    }
 }
