+++ app/src/main/java/com/example/cameraxpractice/PictureViewEdit.kt
@@ -1,0 +1,19 @@
+package com.example.cameraxpractice
+
+import android.graphics.Matrix
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.camera.core.ImageProxy
+import androidx.camera.view.PreviewView
+
+class PictureViewEdit : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_picture_view_edit)
+    }
+//    take in the new picture
+//    display it
+//    set up image matrix
+//    register finger touches for corner
+
+}
