--- app/src/main/java/com/example/cameraxpractice/pictureViewEdit.kt
@@ -1,61 +1,0 @@
-package com.example.cameraxpractice
-
-import android.graphics.Matrix
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.camera.core.ImageProxy
-import androidx.camera.view.PreviewView
-
-class pictureViewEdit : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_picture_view_edit)
-    }
-
-
-
-    fun getCorrectionMatrix(imageProxy: ImageProxy, previewView: PreviewView) : Matrix {
-        val cropRect = imageProxy.cropRect
-        val rotationDegrees = imageProxy.imageInfo.rotationDegrees
-        val matrix = Matrix()
-
-        // A float array of the source vertices (crop rect) in clockwise order.
-        val source = floatArrayOf(
-            cropRect.left.toFloat(),
-            cropRect.top.toFloat(),
-            cropRect.right.toFloat(),
-            cropRect.top.toFloat(),
-            cropRect.right.toFloat(),
-            cropRect.bottom.toFloat(),
-            cropRect.left.toFloat(),
-            cropRect.bottom.toFloat()
-        )
-
-        // A float array of the destination vertices in clockwise order.
-        val destination = floatArrayOf(
-            0f,
-            0f,
-            previewView.width.toFloat(),
-            0f,
-            previewView.width.toFloat(),
-            previewView.height.toFloat(),
-            0f,
-            previewView.height.toFloat()
-        )
-
-        // The destination vertexes need to be shifted based on rotation degrees. The
-        // rotation degree represents the clockwise rotation needed to correct the image.
-
-        // Each vertex is represented by 2 float numbers in the vertices array.
-        val vertexSize = 2
-        // The destination needs to be shifted 1 vertex for every 90° rotation.
-        val shiftOffset = rotationDegrees / 90 * vertexSize;
-        val tempArray = destination.clone()
-        for (toIndex in source.indices) {
-            val fromIndex = (toIndex + shiftOffset) % source.size
-            destination[toIndex] = tempArray[fromIndex]
-        }
-        matrix.setPolyToPoly(source, 0, destination, 0, 4)
-        return matrix
-    }
-}
