+++ app/src/main/res/layout/activity_picture_view_edit.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".pictureViewEdit">
+    tools:context=".PictureViewEdit">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
