+++ build.gradle
@@ -18,6 +18,8 @@ allprojects {
     repositories {
         google()
         jcenter()
+        maven { url "https://jitpack.io" } // for this library
+        maven { url 'https://artifactory.mxtracks.info/artifactory/gradle-dev-local/' } // for the OpenCV library
     }
 }
